Eunomia-package {Eunomia}	R Documentation

Eunomia: A Standard Dataset in the OMOP Common Data Model

Description

A sample dataset in the OMOP (Observational Medical Outcomes Partnership) Common Data Model (CDM) format. The CDM enables uniform storage of observational health care data, and is widely used for health care analytics. 'Eunomia' contains simulated data as well as a subset of the OMOP Vocabulary, and enables testing of additional packages and is used for educational and demonstration purposes.

Author(s)

Maintainer: Frank DeFalco fdefalco@ohdsi.org

Authors:

Martijn Schuemie schuemie@ohdsi.org

Other contributors:

Observational Health Data Science and Informatics [copyright holder]

See Also

Useful links:

https://github.com/OHDSI/Eunomia

Report bugs at https://github.com/OHDSI/Eunomia/issues

[Package Eunomia version 1.0.2 Index]