getEunomiaConnectionDetails {Eunomia}	R Documentation

Get Eunomia Connection Details

Description

Creates a copy of the Eunomia database, and provides details for connecting to that copy.

Usage

getEunomiaConnectionDetails(databaseFile = tempfile(fileext = ".sqlite"))

Arguments

databaseFile
The path where the database file will be copied to. By default, the database will be copied to a temporary folder, and will be deleted at the end of the R session.

Value

A ConnectionDetails object, to be used with the DatabaseConnector package.

Examples

connectionDetails <- getEunomiaConnectionDetails()
connection <- connect(connectionDetails)
querySql(connection, "SELECT COUNT(*) FROM person;")
disconnect(connection)

[Package Eunomia version 1.0.2 Index]